:memo: tree preOrder iterative
