+++ src/main/kotlin/alogs/BFS.kt
@@ -1,18 +1,5 @@
 package alogs
 
-abstract class AbstractGraph<T> {
-    data class Edges<T>(val pointsTo: Vertex<T>)
-    data class Vertex<T>(val data: T, val edges: List<Edges<T>>)
-}
-
-interface Graph<T> {
-    fun breadthTraversal()
-    fun depthTraversal()
-}
-
-interface MutableGraph<T> : Graph<T>
-
-class GraphCollection
 
 fun main() {
